+++ src/main/java/org/cdpg/dx/database/elastic/model/AccessPolicyQueryDecorator.java
@@ -1,74 +1,66 @@
 package org.cdpg.dx.database.elastic.model;
 
-import static org.cdpg.dx.database.elastic.util.Constants.ACCESS_POLICY;
-import static org.cdpg.dx.tgdex.util.Constants.*;
+import static org.cdpg.dx.database.elastic.util.Constants.*;
+import static org.cdpg.dx.tgdex.util.Constants.FIELD;
+import static org.cdpg.dx.tgdex.util.Constants.PROVIDER_USER_ID;
 
-import io.vertx.core.json.JsonObject;
 import java.util.List;
 import java.util.Map;
-import org.cdpg.dx.database.elastic.util.QueryType;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.cdpg.dx.database.elastic.util.*;
 
 public class AccessPolicyQueryDecorator implements ElasticsearchQueryDecorator {
-
+  private static final Logger LOGGER = LogManager.getLogger(AccessPolicyQueryDecorator.class);
   private final Map<FilterType, List<QueryModel>> queryMap;
-  private final JsonObject request;
+  private final AccessPolicyRequestDTO request;
 
-  public AccessPolicyQueryDecorator(Map<FilterType, List<QueryModel>> queryMap,
-                                    JsonObject request) {
+  public AccessPolicyQueryDecorator(
+      Map<FilterType, List<QueryModel>> queryMap, AccessPolicyRequestDTO request) {
     this.queryMap = queryMap;
-      this.request = request;
+    this.request = request;
   }
 
   @Override
   public Map<FilterType, List<QueryModel>> add() {
-    String sub = request.getString(SUB);
-    boolean isMyAssetsRequest = request.getBoolean(MY_ASSETS_REQ, false);
+    LOGGER.info("Adding access policy query decorator DTO {}", request);
+    String sub = request.getSub();
+    boolean isMyAssetsRequest = Boolean.TRUE.equals(request.getMyAssetsReq());
 
     if (sub != null && !sub.isEmpty()) {
       if (isMyAssetsRequest) {
         // Strictly match owned items only
-        QueryModel ownerMatch = new QueryModel(QueryType.MATCH)
+        QueryModel ownerMatch =
+            new QueryModel(QueryType.MATCH)
                 .setQueryParameters(Map.of(FIELD, PROVIDER_USER_ID, VALUE, sub));
         queryMap.get(FilterType.MUST).add(ownerMatch);
       } else {
         // User is authenticated → allow: PUBLIC, RESTRICTED, PRIVATE owned
-
-        // public
-        QueryModel publicAccess = new QueryModel(QueryType.MATCH)
+        QueryModel publicAccess =
+            new QueryModel(QueryType.MATCH)
                 .setQueryParameters(Map.of(FIELD, ACCESS_POLICY, VALUE, OPEN));
-
-        // restricted
-        QueryModel restrictedAccess = new QueryModel(QueryType.MATCH)
+        QueryModel restrictedAccess =
+            new QueryModel(QueryType.MATCH)
                 .setQueryParameters(Map.of(FIELD, ACCESS_POLICY, VALUE, RESTRICTED));
-
-        // private
-        QueryModel privateAccess = new QueryModel(QueryType.MATCH)
+        QueryModel privateAccess =
+            new QueryModel(QueryType.MATCH)
                 .setQueryParameters(Map.of(FIELD, ACCESS_POLICY, VALUE, PRIVATE));
-
-        // owner match
-        QueryModel ownerMatch = new QueryModel(QueryType.MATCH)
+        QueryModel ownerMatch =
+            new QueryModel(QueryType.MATCH)
                 .setQueryParameters(Map.of(FIELD, PROVIDER_USER_ID, VALUE, sub));
-
-        // private AND owned
-        QueryModel privateOwned = new QueryModel(QueryType.BOOL)
-                .setMustQueries(List.of(privateAccess, ownerMatch));
-
-        // OR clause: public OR restricted OR private+owned
+        QueryModel privateOwned =
+            new QueryModel(QueryType.BOOL).setMustQueries(List.of(privateAccess, ownerMatch));
         QueryModel accessFilter = new QueryModel(QueryType.BOOL);
         accessFilter.setShouldQueries(List.of(publicAccess, restrictedAccess, privateOwned));
         accessFilter.setMinimumShouldMatch("1");
-
         queryMap.get(FilterType.MUST).add(accessFilter);
-
       }
     } else {
-      // Not authenticated → exclude PRIVATE
-      QueryModel excludePrivate = new QueryModel(QueryType.MATCH)
+      QueryModel excludePrivate =
+          new QueryModel(QueryType.MATCH)
               .setQueryParameters(Map.of(FIELD, ACCESS_POLICY, VALUE, PRIVATE));
-
       queryMap.get(FilterType.MUST_NOT).add(excludePrivate);
     }
-
     return queryMap;
   }
 }
