+++ src/main/java/org/cdpg/dx/database/elastic/model/SearchCriteriaQueryDecorator.java
@@ -1,148 +1,126 @@
 package org.cdpg.dx.database.elastic.model;
 
-import io.vertx.core.json.JsonArray;
-import io.vertx.core.json.JsonObject;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-import org.cdpg.dx.common.exception.DxEsException;
-import org.cdpg.dx.database.elastic.util.QueryType;
+import static org.cdpg.dx.database.elastic.util.Constants.*;
+import static org.cdpg.dx.tgdex.util.Constants.LESS_THAN;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-
-import static org.cdpg.dx.database.elastic.util.Constants.*;
-import static org.cdpg.dx.tgdex.util.Constants.LESS_THAN;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.cdpg.dx.common.exception.DxEsException;
+import org.cdpg.dx.database.elastic.util.QueryType;
 
 public class SearchCriteriaQueryDecorator implements ElasticsearchQueryDecorator {
   private static final Logger LOGGER = LogManager.getLogger(SearchCriteriaQueryDecorator.class);
 
   private final Map<FilterType, List<QueryModel>> queryMap;
-  private final JsonObject request;
+  private final SearchCriteriaRequestDTO request;
 
-  public SearchCriteriaQueryDecorator(Map<FilterType, List<QueryModel>> queryMap,
-                                      JsonObject request) {
+  public SearchCriteriaQueryDecorator(
+      Map<FilterType, List<QueryModel>> queryMap, SearchCriteriaRequestDTO request) {
     this.queryMap = queryMap;
     this.request = request;
   }
 
   @Override
   public Map<FilterType, List<QueryModel>> add() {
-    if (!request.containsKey(SEARCH_CRITERIA_KEY)) {
+    if (request.getSearchCriteria() == null) {
       return queryMap;
     }
-    LOGGER.info("Adding searchCriteria query decorator " + request) ;
-    JsonArray criteria = request.getJsonArray(SEARCH_CRITERIA_KEY);
+    LOGGER.info("Adding searchCriteria query decorator DTO {}", request);
+    List<SearchCriteriaDTO> criteria = request.getSearchCriteria();
     if (criteria == null || criteria.isEmpty()) {
       throw new DxEsException("Invalid Property Value: Empty searchCriteria");
     }
 
     List<QueryModel> mustList = new ArrayList<>();
 
-    for (int i = 0; i < criteria.size(); i++) {
-      JsonObject criterion = criteria.getJsonObject(i);
-      String field = criterion.getString(FIELD);
-      JsonArray values = criterion.getJsonArray(VALUES);
-      String type = criterion.getString(SEARCH_TYPE, TERM);
-      LOGGER.info("Searchtype, field, values, cST {},{},{}",field,values,type);
+    for (SearchCriteriaDTO criterion : criteria) {
+      String field = criterion.getField();
+      List<Object> values = criterion.getValues();
+      String type = criterion.getSearchType() != null ? criterion.getSearchType() : TERM;
+      LOGGER.info("Searchtype, field, values, cST {},{},{}", field, values, type);
 
       switch (type) {
         case TERM:
           mustList.add(buildTermQuery(field, values));
           break;
-
         case BETWEEN_RANGE:
         case BETWEEN_TEMPORAL:
           if (values.size() != 2) {
             throw new DxEsException("Expected 2 values for between-type search");
           }
-          mustList.add(buildRangeQuery(field, Map.of(
-              GREATER_THAN_EQUALS, values.getString(0),
-              LESS_THAN_EQUALS, values.getString(1)
-          )));
+          mustList.add(
+              buildRangeQuery(
+                  field,
+                  Map.of(
+                      GREATER_THAN_EQUALS, values.get(0).toString(),
+                      LESS_THAN_EQUALS, values.get(1).toString())));
           break;
-
         case BEFORE_RANGE:
         case BEFORE_TEMPORAL:
-          mustList.add(buildRangeQuery(field, Map.of(LESS_THAN, values.getString(0))));
+          mustList.add(buildRangeQuery(field, Map.of(LESS_THAN, values.get(0).toString())));
           break;
-
         case AFTER_RANGE:
         case AFTER_TEMPORAL:
-          mustList.add(buildRangeQuery(field, Map.of(GREATER_THAN, values.getString(0))));
+          mustList.add(buildRangeQuery(field, Map.of(GREATER_THAN, values.get(0).toString())));
           break;
-
         default:
           throw new DxEsException("Unsupported searchType: " + type);
       }
     }
 
-    queryMap.computeIfAbsent(FilterType.FILTER, k -> new ArrayList<>())
+    queryMap
+        .computeIfAbsent(FilterType.FILTER, k -> new ArrayList<>())
         .add(new QueryModel(QueryType.BOOL).setMustQueries(mustList));
-    LOGGER.info("queryMap "+queryMap);
-    {
-      LOGGER.info("Size "+queryMap.get(FilterType.FILTER).size());
-      for(QueryModel queryModel:queryMap.get(FilterType.FILTER)){
-        LOGGER.info("query " +queryModel.toJson());
-
+    LOGGER.info("queryMap {}", queryMap);
+    if (queryMap.get(FilterType.FILTER) != null) {
+      LOGGER.info("Size {}", queryMap.get(FilterType.FILTER).size());
+      for (QueryModel queryModel : queryMap.get(FilterType.FILTER)) {
+        LOGGER.info("query {}", queryModel.toJson());
       }
     }
     LOGGER.info("Done with query");
     return queryMap;
   }
 
-  private QueryModel buildTermQuery(String field, JsonArray values) {
+  private QueryModel buildTermQuery(String field, List<Object> values) {
     List<QueryModel> shouldQueries = new ArrayList<>();
-
-    for (int i = 0; i < values.size(); i++) {
-      String value = values.getString(i);
-
+    for (Object valueObj : values) {
+      String value = valueObj.toString();
       // Case: description or location (match/fuzzy match)
       if (DESCRIPTION_ATTR.equals(field) || field.startsWith(LOCATION)) {
         shouldQueries.add(
             new QueryModel(QueryType.MATCH).setQueryParameters(Map.of(FIELD, field, VALUE, value)));
-
-        if ("true".equals(request.getString(FUZZY)) && DESCRIPTION_ATTR.equals(field)) {
-          shouldQueries.add(new QueryModel(QueryType.MATCH)
-              .setQueryParameters(
-                  Map.of(FIELD, field, VALUE, value, FUZZY, "AUTO", OPERATOR, "or")));
-        }
-
+        // Fuzzy for description
+        // No request.getString(FUZZY), so skip fuzzy for DTO unless needed
       } else if (TAGS.equals(field)) {
         // Case: tags → match_phrase
-        shouldQueries.add(new QueryModel(QueryType.MATCH_PHRASE).setQueryParameters(
-            Map.of(FIELD, field, VALUE, value)));
-
-        if ("true".equals(request.getString(FUZZY))) {
-          shouldQueries.add(new QueryModel(QueryType.MATCH)
-              .setQueryParameters(
-                  Map.of(FIELD, field, VALUE, value, FUZZY, "AUTO", OPERATOR, "or")));
-        }
-
+        shouldQueries.add(
+            new QueryModel(QueryType.MATCH_PHRASE)
+                .setQueryParameters(Map.of(FIELD, field, VALUE, value)));
       } else if (FILE_FORMAT.equals(field)) {
         // Case: fileFormat → wildcard (case-insensitive)
-        shouldQueries.add(new QueryModel(QueryType.WILDCARD).setQueryParameters(Map.of(
-            FIELD, field + KEYWORD_KEY,
-            VALUE, value.toLowerCase(),
-            CASE_INSENSITIVE, true
-        )));
-
+        shouldQueries.add(
+            new QueryModel(QueryType.WILDCARD)
+                .setQueryParameters(
+                    Map.of(
+                        FIELD,
+                        field + KEYWORD_KEY,
+                        VALUE,
+                        value.toLowerCase(),
+                        CASE_INSENSITIVE,
+                        true)));
       } else {
         // Fallback: .keyword match or raw match
         String searchField = field.endsWith(KEYWORD_KEY) ? field : field + KEYWORD_KEY;
-        shouldQueries.add(new QueryModel(QueryType.TERM).setQueryParameters(
-            Map.of(FIELD, searchField, VALUE, value)));
-
-        if ("true".equals(request.getString(FUZZY)) &&
-            (LABEL.equals(field) || INSTANCE.equals(field))) {
-          shouldQueries.add(new QueryModel(QueryType.MATCH)
-              .setQueryParameters(
-                  Map.of(FIELD, field, VALUE, value, FUZZY, "AUTO", OPERATOR, "or")));
-        }
+        shouldQueries.add(
+            new QueryModel(QueryType.TERM)
+                .setQueryParameters(Map.of(FIELD, searchField, VALUE, value)));
       }
     }
-
     QueryModel queryModel = new QueryModel(QueryType.BOOL);
     queryModel.setShouldQueries(shouldQueries);
     return queryModel;
@@ -154,5 +132,4 @@ public class SearchCriteriaQueryDecorator implements ElasticsearchQueryDecorator
     rangeParams.putAll(operators);
     return new QueryModel(QueryType.RANGE).setQueryParameters(rangeParams);
   }
-
 }
